requirements.txt
pyopengl>=3.1.9
pyopengl-accelerate>=3.1.9
numpy>=2.2.3
